backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
yt-dlp>=2023.10.0
gunicorn>=21.2.0
